Move Dofnum to a separate location
