+++ skfem/assembly/basis/basis.py
@@ -6,7 +6,7 @@ import numpy as np
 from numpy import ndarray
 
 from skfem.assembly.dofs import Dofs
-from skfem.assembly.basis.dofnum import Dofnum
+from skfem.assembly.dofnum import Dofnum
 from skfem.element.discrete_field import DiscreteField
 from skfem.element.element_composite import ElementComposite
