+++ skfem/assembly/dofnum.py
