+++ skfem/mesh/geometry.py
@@ -1,4 +1,3 @@
-import numpy as np
 from numpy import ndarray
 
 from skfem import MappingIsoparametric
